feat: refactor handlers into their own directory
